feat: add migration leaderboardEntry, model leaderboardEntry, and controller leaderboardEntry

+++ database/migrations/2024_11_25_044417_create_quizzes_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->integer('exp_received');
             $table->integer('total_question');
             $table->integer('quiz_time');
-            $table->unsignedBigInteger('user_id');
+            $table->Integer('user_id');
             $table->foreign('user_id')->references('user_id')->on('users')->onDelete('cascade');
             $table->integer('correct_question');
             $table->timestamps();
